fix(libtl): typecast expression to match data type
This corrects the MISRA violation C2012-10.1:
Operands shall not be of an inappropriate essential type.
The original code used an expression of a non-Boolean essential type
in a Boolean context. To comply with the rule, the expression has been
replaced with an explicit comparison to a Boolean-compatible
value (e.g., == NULL, != 0U, ensuring clarity and type safety.

Change-Id: Iac29e26415267e6287b523e1535fcda1d4fcf0de
Signed-off-by: Pranav Tilak <pranav.vinaytilak@amd.com>